backend/requirements.txt
# Anime Voice Forge — Backend Dependencies
# Python 3.11+ required

coqui-tts==0.27.5
torch==2.4.1
torchaudio==2.4.1
numpy==1.26.4
transformers==4.57.6
flask==3.1.3
